Migrate min(dim=?) from THC to ATen
