+++ aten/src/ATen/NumericUtils.h
@@ -43,7 +43,7 @@ inline bool _isnan(T val) {
 template <typename T,
          typename std::enable_if<std::is_same<T, at::Half>::value, int>::type = 0>
 inline bool _isnan(T val) {
-  return true;
+  return at::_isnan(float(val));
 }
