+++ aten/src/ATen/native/ReduceOps.cpp
@@ -38,154 +38,6 @@ DEFINE_DISPATCH(argmin_stub);
 DEFINE_DISPATCH(cumsum_stub);
 DEFINE_DISPATCH(cumprod_stub);
 
-#define OPTION_TYPE_EQUALITY_CHECK(option, out, self) \
-{ \
-  TORCH_CHECK(\
-    out.option() == self.option(),\
-    "expected ", #option, " ",\
-    self.option(),\
-    " but found ", out.option())\
-}
-
-static inline void check_scalar_type_device_layout_equal(const Tensor& out, const Tensor& self) {
-  OPTION_TYPE_EQUALITY_CHECK(scalar_type, out, self);
-  OPTION_TYPE_EQUALITY_CHECK(device, out.options(), self.options());
-  OPTION_TYPE_EQUALITY_CHECK(layout, out.options(), self.options());
-}
-
-static inline Tensor integer_upcast(const Tensor& self, optional<ScalarType> dtype) {
-  ScalarType scalarType = self.scalar_type();
-  ScalarType upcast_scalarType = dtype.value_or(at::isIntegralType(scalarType, /*includeBool=*/true) ? ScalarType::Long : scalarType);
-  return self.toType(upcast_scalarType);
-}
-
-using DimMask = TensorIterator::DimMask;
-
-static DimMask make_dim_mask(IntArrayRef dims, int64_t ndim) {
-  auto mask = DimMask();
-  if (dims.empty()) {
-    mask.flip();
-  } else {
-    for (int64_t dim : dims) {
-      int64_t pos_dim = maybe_wrap_dim(dim, ndim);
-      TORCH_CHECK(pos_dim < 64, "PyTorch doesn't support reduction operations for dim>=64");
-      mask.set(pos_dim);
-    }
-  }
-  return mask;
-}
-
-static void allocate_reduction_result(
-    Tensor& result, const Tensor& self, DimMask mask, bool keepdim,
-    ScalarType dtype)
-{
-  auto shape = DimVector(self.sizes());
-  for (int dim = shape.size() - 1; dim >= 0; dim--) {
-    if (mask[dim]) {
-      if (keepdim) {
-        shape[dim] = 1;
-      } else {
-        shape.erase(shape.begin() + dim);
-      }
-    }
-  }
-  if (result.defined()) {
-    result.resize_(shape);
-  } else {
-    result = at::empty(shape, self.options().dtype(dtype));
-  }
-}
-
-static Tensor review_reduce_result(const Tensor& result, int ndim, DimMask mask, bool keepdim) {
-  if (keepdim) {
-    return result;
-  }
-  auto shape = DimVector(result.sizes());
-  auto stride = DimVector(result.strides());
-  for (int dim = 0; dim < ndim; dim++) {
-    if (mask[dim]) {
-      shape.insert(shape.begin() + dim, 1);
-      stride.insert(stride.begin() + dim, 0);
-    }
-  }
-  return result.as_strided(shape, stride);
-}
-
-static TensorIterator make_reduction(
-    const char* name, Tensor& result, const Tensor& self, IntArrayRef dim,
-    bool keepdim, ScalarType in_dtype, ScalarType out_dtype)
-{
-  // check that result type and dtype match if provided
-  TORCH_CHECK(
-      !result.defined() || result.scalar_type() == out_dtype,
-      name, ": provided dtype must match dtype of result. Got ",
-      toString(result.scalar_type()),
-      " and ",
-      toString(out_dtype),
-      ".");
-  int64_t ndim = self.dim();
-  auto mask = make_dim_mask(dim, ndim);
-  allocate_reduction_result(result, self, mask, keepdim, out_dtype);
-  auto viewed_result = review_reduce_result(result, ndim, mask, keepdim);
-  namedinference::propagate_names_for_reduction(result, self, dim, keepdim);
-  if (self.scalar_type() == in_dtype) {
-    return TensorIterator::reduce_op(viewed_result, self);
-  }
-  return TensorIterator::reduce_op(viewed_result, self.to(in_dtype));
-}
-
-static TensorIterator make_reduction(
-    const char* name, Tensor& result, const Tensor& self, IntArrayRef dim,
-    bool keepdim, ScalarType out_dtype)
-{
-  // special case for type promotion in mixed precision, improves computational
-  // efficiency.
-  // not generalize this to common mismatched input/output types to avoid cross
-  // product of templated kernel launches.
-  const bool gpu_f16_to_f32 = (
-    self.is_cuda() && self.scalar_type() == kHalf && out_dtype == kFloat);
-  auto in_dtype = gpu_f16_to_f32 ? self.scalar_type() : out_dtype;
-  return make_reduction(name, result, self, dim, keepdim, in_dtype, out_dtype);
-}
-
-static TensorIterator make_reduction(
-    const char* name, Tensor& result1, Tensor& result2, const Tensor& self, IntArrayRef dim,
-    bool keepdim, ScalarType dtype)
-{
-  // check that result type and dtype match if provided
-  for (const Tensor *t: {&result1, &result2}) {
-    const Tensor& result = *t;
-    TORCH_CHECK(
-        !result.defined() || result.scalar_type() == dtype,
-        name, ": provided dtype must match dtype of result. Got ",
-        toString(result.scalar_type()),
-        " and ",
-        toString(dtype),
-        ".");
-  }
-
-  int64_t ndim = self.dim();
-  DimMask mask = make_dim_mask(dim, ndim);
-  allocate_reduction_result(result1, self, mask, keepdim, dtype);
-  auto viewed_result1 = review_reduce_result(result1, ndim, mask, keepdim);
-
-  allocate_reduction_result(result2, self, mask, keepdim, dtype);
-  auto viewed_result2 = review_reduce_result(result2, ndim, mask, keepdim);
-
-  namedinference::propagate_names_for_reduction(result1, self, dim, keepdim);
-  namedinference::propagate_names_for_reduction(result2, self, dim, keepdim);
-
-  // special case for type promotion in mixed precision, improves computational
-  // efficiency.
-  // We don't generalize this to common mismatched input/output types to avoid cross
-  // product of templated kernel launches.
-  if (self.scalar_type() == dtype ||
-      (self.is_cuda() && self.scalar_type() == kHalf && dtype == kFloat)) {
-    return TensorIterator::reduce_op(viewed_result1, viewed_result2, self);
-  }
-  return TensorIterator::reduce_op(viewed_result1, viewed_result2, self.to(dtype));
-}
-
 Tensor _cumsum_cpu(const Tensor& self, int64_t dim) {
   Tensor result = at::empty_like(self, MemoryFormat::Contiguous);
   cumsum_stub(self.device().type(), result, self, dim);
