+++ aten/src/ATen/native/TensorCompare.cpp
@@ -259,18 +259,18 @@ std::tuple<Tensor&,Tensor&> max_out(Tensor& max, Tensor& max_indices,
   return result;
 }
 
-std::tuple<Tensor &,Tensor &> _min_out_cpu(Tensor& min, Tensor& min_indices,
+std::tuple<Tensor &,Tensor &> _min_out(Tensor& min, Tensor& min_indices,
                                         const Tensor& self, int64_t dim, bool keepdim) {
   TORCH_CHECK(!self.is_complex(), "min is not yet implemented for complex tensors.");
-  min_stub(kCPU, min, min_indices, self, dim, keepdim);
+  min_stub(self.device().type(), min, min_indices, self, dim, keepdim);
   return std::tuple<Tensor &,Tensor &>{min, min_indices};
 }
 
-std::tuple<Tensor, Tensor> _min_cpu(const Tensor& self, int64_t dim, bool keepdim) {
+std::tuple<Tensor, Tensor> _min(const Tensor& self, int64_t dim, bool keepdim) {
   TORCH_CHECK(!self.is_complex(), "min is not yet implemented for complex tensors.");
-  Tensor min_indices = at::empty({0}, self.options().dtype(kLong));
-  Tensor min = at::empty({0}, self.options());
-  return at::native::_min_out_cpu(min, min_indices, self, dim, keepdim);
+  Tensor min_indices;
+  Tensor min;
+  return at::native::_min_out(min, min_indices, self, dim, keepdim);
 }
 
 std::tuple<Tensor, Tensor> min(const Tensor& self, int64_t dim, bool keepdim) {
