+++ aten/src/ATen/native/TensorCompare.h
@@ -7,11 +7,11 @@
 
 namespace at { namespace native {
 
-using reduce_fn =
+using reduce_minmax_fn =
     void (*)(Tensor&, Tensor&, const Tensor&, int64_t, bool);
 
-DECLARE_DISPATCH(reduce_fn, max_stub);
-DECLARE_DISPATCH(reduce_fn, min_stub);
+DECLARE_DISPATCH(reduce_minmax_fn, max_stub);
+DECLARE_DISPATCH(reduce_minmax_fn, min_stub);
 
 using where_fn = void (*)(TensorIterator &, ScalarType);
 DECLARE_DISPATCH(where_fn, where_kernel);
