+++ aten/src/ATen/native/native_functions.yaml
@@ -5599,14 +5599,8 @@
 
 - func: _min(Tensor self, int dim, bool keepdim=False) -> (Tensor, Tensor)
   use_c10_dispatcher: full
-  dispatch:
-    CPU: _min_cpu
-    CUDA: legacy::cuda::_th_min
 
-- func: _min.min(Tensor self, int dim, bool keepdim=False, *, Tensor(a!) min, Tensor(b!) min_indices) -> (Tensor(a!), Tensor(b!))
-  dispatch:
-    CPU: _min_out_cpu
-    CUDA: legacy::cuda::_th_min_out
+- func: _min.out(Tensor self, int dim, bool keepdim=False, *, Tensor(a!) min, Tensor(b!) min_indices) -> (Tensor(a!), Tensor(b!))
 
 - func: bucketize.Tensor(Tensor self, Tensor boundaries, *, bool out_int32=False, bool right=False) -> Tensor
   use_c10_dispatcher: full
